+++ modules/random/rand.go
@@ -7,7 +7,7 @@ import (
 
 type DocTxMsgRequestRand struct {
 	Consumer      string        `bson:"consumer"`       // request address
-	BlockInterval uint64        `bson:"block_interval"` // block interval after which the requested random number will be generated
+	BlockInterval int64         `bson:"block_interval"` // block interval after which the requested random number will be generated
 	Oracle        bool          `bson:"oracle"`
 	ServiceFeeCap []models.Coin `bson:"service_fee_cap"`
 }
@@ -19,7 +19,7 @@ func (doctx *DocTxMsgRequestRand) GetType() string {
 func (doctx *DocTxMsgRequestRand) BuildMsg(txMsg interface{}) {
 	msg := txMsg.(*MsgRequestRandom)
 	doctx.Consumer = msg.Consumer
-	doctx.BlockInterval = msg.BlockInterval
+	doctx.BlockInterval = int64(msg.BlockInterval)
 	doctx.Oracle = msg.Oracle
 	doctx.ServiceFeeCap = models.BuildDocCoins(msg.ServiceFeeCap)
 }
